+++ components/description.component.jsx
@@ -15,7 +15,7 @@ const Description = ({ data, repoData }) => {
   )
   return (
     //temp margin
-    <section className="mb-20 p-6 text-white/80 flex gap-10 flex-col justify-between bg-lightPurple border-[1px] border-white/10 rounded-lg md:p-10">
+    <section className="p-6 text-white/80 flex gap-10 flex-col justify-between bg-lightPurple border-[1px] border-white/10 rounded-lg md:p-10">
       <div className="flex gap-6 flex-col justify-between w-full md:gap-10 md:flex-row">
         <div className="flex flex-col justify-start items-start gap-2">
           <h3 className="gradient-text text-xl font-medium">Profile Bio</h3>
