+++ components/footer.component.jsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <footer className='text-center text-xs pt-10 pb-6 text-white/70 md:text-sm'>
+      <div className="container flex items-center justify-center gap-2">
+        <p>Built by </p>
+        <div className="gradient-text flex gap-[2px] items-center">
+          <p className='text-[14px] md:text-[20px] -mb-[2px]'>&copy;</p>
+          <p className="gradient-text ">Zelfroster </p>
+        </div>
+        <p>{Date().split(" ").at(3)}</p>
+      </div>
+    </footer>
+  )
+}
+
+export default Footer
